+++ Cargo.toml
@@ -88,8 +88,8 @@ secp256k1 = { version = "0.29.1", default-features = false, features = [
 ] }
 serde = { version = "1.0", default-features = false }
 serde_json = "1.0"
-sysinfo = "0.37"
 syn = { version = "2", features = ["full"] }
+sysinfo = "0.37"
 thiserror = "2"
 tiny-keccak = "2.0.0"
 tokio = { version = "1", default-features = false }
