+++ crates/node/src/lib.rs
@@ -1,8 +1,8 @@
-pub mod block_queue;
 pub mod error;
 pub mod initializers;
 pub mod node;
 pub mod p2p;
+pub mod pending_heap;
 pub mod rpc;
 pub mod services;
 pub mod types;
