+++ crates/node/src/rpc/context.rs
@@ -1,7 +1,7 @@
-use crate::block_queue::{OrderedBlock, PendingHeap};
+use crate::pending_heap::PendingHeap;
 use ethrex_rpc::{EthClient, RpcApiContext as L1Context};
 use ethrex_storage_rollup::StoreRollup;
-use mojave_utils::unique_heap::AsyncUniqueHeap;
+use mojave_utils::{ordered_block::OrderedBlock, unique_heap::AsyncUniqueHeap};
 
 #[derive(Clone, Debug)]
 pub struct RpcApiContext {
