+++ crates/utils/src/lib.rs
@@ -3,6 +3,7 @@ pub mod convert;
 pub mod daemon;
 pub mod logging;
 pub mod network;
+pub mod ordered_block;
 pub mod p2p;
 pub mod rpc;
 pub mod unique_heap;
